feat: update header to 1.21.60

+++ src/mc/entity/components_json_legacy/MutationStrategy.h
@@ -2,7 +2,7 @@
 
 #include "mc/_HeaderOutputPredefine.h"
 
-enum class MutationStrategy : int {
+enum class MutationStrategy : uchar {
     None   = 0,
     Random = 1,
 };
