+++ src/mc/world/actor/animation/BoneAnimationRelativeMode.h
@@ -2,7 +2,7 @@
 
 #include "mc/_HeaderOutputPredefine.h"
 
-enum class BoneAnimationRelativeMode : int {
+enum class BoneAnimationRelativeMode : uchar {
     Parent = 0,
     Entity = 1,
 };
